pyproject.toml
[project]
name = "rag-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Mrudhuhas", email = "mrudhuhas@outlook.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "google-genai>=1.45.0",
    "llama-index>=0.10.68",
    "llama-index-embeddings-gemini>=0.1.8",
    "llama-index-embeddings-openai>=0.1.11",
    "loguru>=0.7.3",
    "openai>=2.5.0",
    "pinecone[asyncio]>=7.3.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.26.5",
    "redis[hiredis]>=6.4.0",
    "sentence-transformers>=5.1.1",
    "tenacity>=8.2.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.66.0",
]

[project.scripts]
rag-app = "rag_app:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "ipykernel>=7.0.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]
